+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Expose jmespath for import by plugins
+- Implement "computed" group/location relationships
 
 ### Changed